requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
efficientnet_pytorch==0.7.1
fastapi==0.100.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
numpy==2.2.0
opencv_python==4.10.0
packaging==24.2
pillow==11.0.0
pretrainedmodels==0.7.4
pydantic==2.10.3
pydantic_core==2.27.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
segmentation-models-pytorch==0.3.4
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
timm==0.9.7
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.22.0
